requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
python-docx>=1.1.0
extract-msg>=0.48.0
pandas>=2.0.0
xlrd>=2.0.0
